Remove all icons that are not needed in library and move to the sample project. Closes #16
Implement 4 different base themes. Created a new attribute that allows users to specify this theme and the action bar automatically applies these themes and uses the correct icons for up and overflow based on the theme type or if we should use light icons.

Updated the base activity and the other activity to show off the different themes.